Improve description of the tool output
The description of the tool output just says that a greeting will be
displayed. But we shouldn't sell our software short:
We also detect the OS and the JDK version and output them to the user,
which can be really helpful to potential users.

This commit adds a half-sentence to the description to highlight this
fact.